Add functionality to import Python modules from within pure C++ code (#2696).
This functionality is necessary for the table-based persistence framework, which needs to import Python modules to ensure the singleton registry of factories is populated before being searched.  It needs to go in the 'base' package (and outside the 'lsst' python package) because it needs to be imported by the lsstimport.py.
This change adds a C++ shared library to the base package, which is perhaps slightly unfortunate, especially as the library name is simply "base" according to our conventions.