More robust determination of RTLD_* constants
Now attempts to find the RTLD_GLOBAL and RTLD_NOW
constants by importing os, DLFCN and ctypes.

For all cases except OS X Python 2 the constants will
be located. RTLD_NOW will not be determined but on this
system we know that it should be 2. Given this, we remove
the lsst64defs.py dynamic code, which will fix the rare
race condition and fallback to a hard-coded RTLD_NOW
value.